+++ plugins/agk_skin_plus/install.inc.php
@@ -1,0 +1,4 @@
+<?php
+
+$REX['ADDON']['install']['agk_skin_plus'] = true;
+