+++ plugins/agk_skin_plus/settings.inc.php
@@ -1,0 +1,4 @@
+<?php
+// --- DYN
+$REX['ADDON']['agk_skin_plus']['liquid_layout'] = 0;
+// --- /DYN